requirements.txt
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
eth-account>=0.9.0
web3>=6.0.0
